requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyMySQL==1.1.1
requests==2.32.3
beautifulsoup4==4.12.3
loguru==0.7.2
Werkzeug==3.0.3
selenium==4.25.0
webdriver-manager==4.0.2
tenacity==9.0.0
